requirements.txt
pillow
numpy
scipy
reedsolo
cryptography
